@@ -1,6 +1,6 @@
 {
   "name": "@solflare-wallet/sdk",
-  "version": "1.2.1",
+  "version": "1.3.0-beta.2",
   "description": "",
   "repository": "https://github.com/solflare-wallet/solflare-sdk",
   "author": "Solflare Developers <developers@solflare.com>",
@@ -24,31 +24,32 @@
     "lib/"
   ],
   "devDependencies": {
-    "@babel/core": "^7.17.9",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "@solana/web3.js": "^1.62.1",
-    "babel-loader": "^8.2.4",
-    "core-js": "^3.21.1",
-    "css-loader": "^5.2.7",
-    "html-webpack-plugin": "^4.5.2",
+    "@babel/core": "7.18.6",
+    "@babel/preset-env": "^7.22.4",
+    "@babel/preset-react": "^7.22.3",
+    "@solana/web3.js": "^1.77.3",
+    "@types/node": "^20.2.5",
+    "babel-loader": "^9.1.2",
+    "core-js": "^3.30.2",
+    "css-loader": "^6.8.1",
+    "html-webpack-plugin": "^5.5.1",
+    "node-polyfill-webpack-plugin": "^2.0.1",
     "npm-run-all": "^4.1.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "regenerator-runtime": "^0.13.9",
-    "style-loader": "^2.0.0",
-    "typescript": "^4.6.3",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.8.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "regenerator-runtime": "^0.13.11",
+    "style-loader": "^3.3.3",
+    "typescript": "^5.1.3",
+    "webpack": "^5.85.1",
+    "webpack-cli": "^5.1.3",
+    "webpack-dev-server": "^4.15.0"
   },
   "peerDependencies": {
-    "@solana/web3.js": "^1.61.0"
+    "@solana/web3.js": "*"
   },
   "dependencies": {
-    "@project-serum/sol-wallet-adapter": "0.2.0",
-    "bs58": "^4.0.1",
-    "eventemitter3": "^4.0.7",
-    "uuid": "^8.3.2"
+    "bs58": "^5.0.0",
+    "eventemitter3": "^5.0.1",
+    "uuid": "^9.0.0"
   }
 }
