@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import Solflare from '../';
-import { Transaction } from '@solana/web3.js';
+import { Transaction, Connection } from '@solana/web3.js';
 
 export default class App extends Component {
 
@@ -8,6 +8,8 @@ export default class App extends Component {
     super(props);
 
     this.solflare = new Solflare();
+
+    this.connection = new Connection('https://fittest-crimson-night.solana-mainnet.discover.quiknode.pro/ce5a19f97c9f96af9d395263ffb2bdba8ea007eb/', 'confirmed');
   }
 
   state = {
@@ -18,8 +20,9 @@ export default class App extends Component {
     this.solflare.detectWallet().then((detected) => console.log('Wallet detected', detected));
   }
 
-  createTransaction = (publicKey) => {
+  createTransaction = async (publicKey) => {
     const transaction = new Transaction();
+    // transaction.recentBlockhash = (await this.connection.getRecentBlockhash()).blockhash;
     transaction.recentBlockhash = '8sUFDJ1rRV478F2ExDLWpzrvY7Pr6LyM6K2kw7ipzmoS';
     transaction.feePayer = publicKey;
     return transaction;
@@ -59,7 +62,7 @@ export default class App extends Component {
 
   handleSignTransaction = async () => {
     try {
-      const tx = await this.solflare.signTransaction(this.createTransaction(this.solflare.publicKey));
+      const tx = await this.solflare.signTransaction(await this.createTransaction(this.solflare.publicKey));
 
       document.body.append(JSON.stringify(tx));
 
@@ -71,7 +74,7 @@ export default class App extends Component {
 
   handleSignAllTransactions = async () => {
     try {
-      const txs = await this.solflare.signAllTransactions([ this.createTransaction(this.solflare.publicKey), this.createTransaction(this.solflare.publicKey) ]);
+      const txs = await this.solflare.signAllTransactions([ await this.createTransaction(this.solflare.publicKey), await this.createTransaction(this.solflare.publicKey) ]);
 
       document.body.append(JSON.stringify(txs));
 
@@ -93,6 +96,18 @@ export default class App extends Component {
     }
   };
 
+  handleSignAndSendTransaction = async () => {
+    try {
+      const tx = await this.solflare.signAndSendTransaction(await this.createTransaction(this.solflare.publicKey));
+
+      document.body.append(JSON.stringify(tx));
+
+      console.log(tx);
+    } catch (e) {
+      console.log(e);
+    }
+  }
+
   render () {
     if (this.state.connected) {
       return (
@@ -104,6 +119,8 @@ export default class App extends Component {
           <button onClick={this.handleSignMessage}>Sign message</button>
           <br/>
           <button onClick={this.handleDisconnect}>Disconnect</button>
+          <br/>
+          <button onClick={this.handleSignAndSendTransaction}>Sign and send transaction</button>
         </div>
       );
     }
