@@ -1,5 +1,5 @@
 import EventEmitter from 'eventemitter3';
-import { PublicKey } from '@solana/web3.js';
+import { PublicKey, SendOptions } from '@solana/web3.js';
 import { SolflareIframeMessage } from '../types';
 
 export default abstract class WalletAdapter extends EventEmitter {
@@ -10,6 +10,7 @@ export default abstract class WalletAdapter extends EventEmitter {
   abstract disconnect (): Promise<void>;
   abstract signTransaction (message: Uint8Array): Promise<Uint8Array>;
   abstract signAllTransactions (messages: Uint8Array[]): Promise<Uint8Array[]>;
+  abstract signAndSendTransaction (transaction: Uint8Array, options?: SendOptions): Promise<string>;
   abstract signMessage (data: Uint8Array, display: 'hex' | 'utf8'): Promise<Uint8Array>;
   abstract handleMessage (data: SolflareIframeMessage): void;
 }
